Builder remove "object" from doc-block; Cache clean fix

+++ lib/cache/sfFileCache.class.php
@@ -150,7 +150,9 @@ class sfFileCache extends sfCache
     {
       if (sfCache::ALL == $mode || !$this->isValid($file))
       {
-        $result = @unlink($file) && $result;
+          if ($file != '.' && $file != '..' && is_writable($file)) {
+              $result = @unlink($file) && $result;
+          }
       }
     }
